Move test method to existing test
